Make scale slider more usable
Before, the slider was the direct input to the `scale` uniform
in the fragment shader. Due to the inverse relationsip between `scale`
and the width of the output range, this meant that values on the small
end created a very large difference in range.

Now, the slider is a value from 0 to 1 and is put through a quadratic
compensation function to give less weight to lower values, making
navigating low end of the scale more manageable.