Mandelbrot FPS improvements
Unroll complex exponentiation to avoid low framerates with high
iteration counts. In that vein, rather than using the `order` uniform
to dtermine calculation in the fragment shader, a seperate program is
created for each order in range [2,5]. Each shader has it's own
exponentiation hardcoded and unrolled, and the appropriate shader's
`draw` call is used to know which to draw.

This results in between 30 and 120 extra FPS, depending on which
order is chosen. An alternative would have been to limit the number
of iterations but I felt that it was important to get the maximum detail
so that WebGL's floating point precision is the bottleneck for zoom,
not performance.

Unfortunately, this caused me to impose a limit of order 5, but I felt
that higher orders were not as interesting or had enough variety to
justify including them. Maybe in the future I can add an option, but
the additional limitation that GLSL evaluates every branch path before
picking which to draw kinda stinks (like really wth?)